[cli] refactor: extract model/rollout/deployment into platform/cli modules
model, rollout, and deployment implemented their full list (and
deploy/undeploy/info) logic inline in the @app.command bodies, unlike
dataset/train/eval/secret which delegate to a platform/cli/<domain> business
module. deployment.py also embedded a ~50-line interactive checkpoint picker
in the command shell.

Move the logic into platform/cli/{model,rollout,deployment}.py and reduce the
command shells to thin delegators, matching the existing convention. The new
modules adopt the shared paginated_fetch helper from the start. deploy and
undeploy stay importable from cli/commands/deployment.py (registered as
top-level verbs in main.py); the INTERACTIVE_REQUIRED guard moves with the
picker, and the prompt imports stay function-local so source-level patching
keeps working.

Tests patch the business modules directly (OsmosisClient / require_git) to
match the train/secret pattern, removing a reliance on lazy-import timing that
made cross-test FakeClient bindings leak.
