+++ tests/unit/cli/test_deployment_commands.py
@@ -9,8 +9,7 @@ from types import SimpleNamespace
 
 import pytest
 
-import osmosis_ai.cli.commands.deployment as deployment_module
-import osmosis_ai.platform.api.client as api_client_module
+import osmosis_ai.platform.cli.deployment as deployment_module
 import osmosis_ai.platform.cli.utils as utils_module
 from osmosis_ai.cli.console import Console
 from osmosis_ai.cli.errors import CLIError
@@ -64,7 +63,7 @@ def mock_git_context(monkeypatch: pytest.MonkeyPatch) -> SimpleNamespace:
         credentials=AUTH_CREDENTIALS,
     )
     monkeypatch.setattr(
-        "osmosis_ai.platform.cli.utils.require_git_workspace_directory_context",
+        "osmosis_ai.platform.cli.deployment.require_git_workspace_directory_context",
         lambda: context,
     )
     return context
@@ -167,7 +166,7 @@ class TestDeployWizardHelper:
                 return checkpoint
             raise AssertionError(f"unexpected prompt: {message}")
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         monkeypatch.setattr("osmosis_ai.cli.prompts.select_list", fake_select_list)
 
         selected = deployment_module._select_checkpoint_for_deploy(mock_git_context)
@@ -199,7 +198,7 @@ class TestDeployWizardHelper:
             assert actions[0].title == "Cancel"
             return actions[0].value
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         monkeypatch.setattr("osmosis_ai.cli.prompts.select_list", fake_select_list)
 
         assert deployment_module._select_checkpoint_for_deploy(mock_git_context) is None
@@ -244,7 +243,7 @@ class TestDeployWizardHelper:
                 return actions[-1].value
             raise AssertionError(f"unexpected prompt: {message}")
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         monkeypatch.setattr("osmosis_ai.cli.prompts.select_list", fake_select_list)
 
         assert deployment_module._select_checkpoint_for_deploy(mock_git_context) is None
@@ -294,7 +293,7 @@ class TestDeployWizardHelper:
                 return next(selected_checkpoints)
             raise AssertionError(f"unexpected prompt: {message}")
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         monkeypatch.setattr("osmosis_ai.cli.prompts.select_list", fake_select_list)
 
         selected = deployment_module._select_checkpoint_for_deploy(mock_git_context)
@@ -316,7 +315,7 @@ class TestDeployWizardHelper:
         def fake_select_list(message, *, items, actions):
             raise AssertionError("select_list should not be called")
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         monkeypatch.setattr("osmosis_ai.cli.prompts.select_list", fake_select_list)
 
         with pytest.raises(CLIError, match="No training runs"):
@@ -376,7 +375,7 @@ class TestDeployWizardHelper:
             assert default is True
             return True
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         monkeypatch.setattr("osmosis_ai.cli.prompts.select_list", fake_select_list)
         monkeypatch.setattr("osmosis_ai.cli.prompts.confirm", fake_confirm)
 
@@ -421,7 +420,7 @@ class TestDeployWizardHelper:
             assert items
             return run
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         monkeypatch.setattr("osmosis_ai.cli.prompts.select_list", fake_select_list)
         monkeypatch.setattr("osmosis_ai.cli.prompts.confirm", lambda _message: False)
 
@@ -447,7 +446,7 @@ class TestListDeployments:
                     deployments=[], total_count=0, has_more=False
                 )
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         result = deployment_module.list_deployments(limit=30, all_=False)
 
         assert isinstance(result, ListResult)
@@ -482,7 +481,7 @@ class TestListDeployments:
                     deployments=[dep], total_count=1, has_more=False
                 )
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         result = deployment_module.list_deployments(limit=10, all_=False)
         assert captured == {
             "limit": 10,
@@ -520,7 +519,7 @@ class TestInfo:
                     created_at="2026-04-20T00:00:00Z",
                 )
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         result = deployment_module.info(checkpoint=checkpoint_id)
         assert captured == {
             "checkpoint": checkpoint_id,
@@ -548,7 +547,7 @@ class TestInfo:
                     base_model="Qwen/Qwen3",
                 )
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
 
         result = deployment_module.info(checkpoint="qwen3-run1-step-100")
 
@@ -576,7 +575,7 @@ class TestDeploy:
                     status="active",
                 )
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         result = deployment_module.deploy(checkpoint="qwen3-run1-step-100")
         assert captured == {
             "checkpoint": "qwen3-run1-step-100",
@@ -602,7 +601,7 @@ class TestDeploy:
             def deploy_checkpoint(self, checkpoint, *, git_identity, credentials=None):
                 raise AssertionError("checkpoint should not be deployed")
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         monkeypatch.setattr(
             deployment_module,
             "_select_checkpoint_for_deploy",
@@ -642,7 +641,7 @@ class TestDeploy:
                     status="active",
                 )
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
 
         with override_output_context(
             format=OutputFormat.rich, interactive=True
@@ -667,7 +666,7 @@ class TestDeploy:
                     status="failed",
                 )
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         result = deployment_module.deploy(checkpoint="qwen3-run1-step-100")
 
         assert isinstance(result, OperationResult)
@@ -697,7 +696,7 @@ class TestUndeploy:
                     status="inactive",
                 )
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         result = deployment_module.undeploy(checkpoint=checkpoint_id)
         assert captured == {
             "checkpoint": checkpoint_id,
@@ -738,7 +737,7 @@ class TestUndeploy:
                     status="inactive",
                 )
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
 
         with override_output_context(
             format=OutputFormat.rich, interactive=True
@@ -765,7 +764,7 @@ class TestUndeploy:
                     status="failed",
                 )
 
-        monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(deployment_module, "OsmosisClient", FakeClient)
         result = deployment_module.undeploy(checkpoint="qwen3-run1-step-100")
 
         assert isinstance(result, OperationResult)
